Stop movement when robot lifted
This change sets a variable to disable movement when the Turtlebot's
wheels drop (if it falls or is lifted). This change designed to kill
the thread that called move since that we don't want to be hitting the
interrupt button after this. A function is inclued that should be called
to reset movement.